[K6.0] Move remaining methods into Kunenagrid class into service
